+++ fort-to-nite/AppDelegate.swift
@@ -84,10 +84,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                 tabBar.selectedIndex = 2
                 
                 guard let nvc = tabBar.selectedViewController as? UINavigationController else { return false }
-                guard let vc = nvc.viewControllers.first as? MoreTableViewController else { return false }
+                guard let vc = nvc.viewControllers.first as? FavoritesTableViewController else { return false }
                 WeaponViewController().getData()
                 nvc.popToRootViewController(animated: true)
-                vc.performSegue(withIdentifier: "moreToFav", sender: nil)
                 
                 quickActionHandled = true
             }