+++ fort-to-nite/Viewcontrollers/WeaponViewController.swift
@@ -26,7 +26,7 @@ extension UserDefaults {
     }
     
     static func lastUpdate() -> Bool {
-        let lastVersion = "1.0.5"
+        let lastVersion = "1.0.6"
         let actualVersion = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String
         let isANewVersion = !UserDefaults.standard.bool(forKey: lastVersion)